+++ example/lib/main.dart
@@ -2,9 +2,9 @@ import 'package:draggable_scrollbar/draggable_scrollbar.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-void main() => runApp(DragScrollBarDemo());
+void main() => runApp(DraggableScrollBarDemo());
 
-class DragScrollBarDemo extends StatelessWidget {
+class DraggableScrollBarDemo extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -24,41 +24,147 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  ScrollController _controller = ScrollController();
+  ScrollController _googlePhotosController = ScrollController();
+  ScrollController _arrowsController = ScrollController();
+  ScrollController _rrectController = ScrollController();
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
+    return DefaultTabController(
+      length: 3,
+      child: Scaffold(
+        appBar: AppBar(
+          title: Text(widget.title),
+          bottom: TabBar(tabs: [
+            Tab(
+              text: 'Google Photos',
+            ),
+            Tab(
+              text: 'Arrows',
+            ),
+            Tab(
+              text: 'RRect',
+            ),
+          ]),
+        ),
+        body: TabBarView(children: [
+          GooglePhotosDemo(controller: _googlePhotosController),
+          ArrowsDemo(controller: _arrowsController),
+          RRectDemo(controller: _rrectController),
+        ]),
       ),
-      body: Column(
-        children: <Widget>[
-          Expanded(
-            child: DraggableScrollbar.asGooglePhotos(
-              dynamicLabelTextBuilder: (double offset) => "${offset ~/ 100}",
-              controller: _controller,
-              child: ListView.builder(
-                controller: _controller,
-                itemCount: 1000,
-                itemExtent: 100.0,
-                itemBuilder: (context, index) {
-                  return Container(
-                    padding: EdgeInsets.all(10.0),
-                    child: Material(
-                      elevation: 4.0,
-                      borderRadius: BorderRadius.circular(5.0),
-                      color: Colors.purple[index % 9 * 100],
-                      child: Center(
-                        child: Text(index.toString()),
-                      ),
-                    ),
-                  );
-                },
+    );
+  }
+}
+
+class GooglePhotosDemo extends StatelessWidget {
+  static int numItems = 1000;
+
+  final ScrollController controller;
+
+  const GooglePhotosDemo({
+    Key key,
+    @required this.controller,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: <Widget>[
+        Expanded(
+          child: DraggableScrollbar.asGooglePhotos(
+            labelTextBuilder: (offset) {
+              final int currentItem = controller.hasClients
+                  ? (controller.offset /
+                          controller.position.maxScrollExtent *
+                          numItems)
+                      .floor()
+                  : 0;
+
+              return Text("$currentItem");
+            },
+            controller: controller,
+            child: GridView.builder(
+              gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+                crossAxisCount: 5,
               ),
+              controller: controller,
+              padding: EdgeInsets.zero,
+              itemCount: numItems,
+              itemBuilder: (context, index) {
+                return Container(
+                  alignment: Alignment.center,
+                  margin: EdgeInsets.all(2.0),
+                  color: Colors.grey[300],
+                  child: Text('$index'),
+                );
+              },
             ),
           ),
-        ],
+        ),
+      ],
+    );
+  }
+}
+
+class ArrowsDemo extends StatelessWidget {
+  final ScrollController controller;
+
+  const ArrowsDemo({Key key, @required this.controller}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return DraggableScrollbar.withArrows(
+      labelTextBuilder: (double offset) => Text("${offset ~/ 100}"),
+      controller: controller,
+      child: ListView.builder(
+        controller: controller,
+        itemCount: 1000,
+        itemExtent: 100.0,
+        itemBuilder: (context, index) {
+          return Container(
+            padding: EdgeInsets.all(8.0),
+            child: Material(
+              elevation: 4.0,
+              borderRadius: BorderRadius.circular(4.0),
+              color: Colors.purple[index % 9 * 100],
+              child: Center(
+                child: Text(index.toString()),
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
+
+class RRectDemo extends StatelessWidget {
+  final ScrollController controller;
+
+  const RRectDemo({Key key, @required this.controller}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return DraggableScrollbar.rrect(
+      controller: controller,
+      child: ListView.builder(
+        controller: controller,
+        itemCount: 1000,
+        itemExtent: 100.0,
+        itemBuilder: (context, index) {
+          return Container(
+            padding: EdgeInsets.all(8.0),
+            child: Material(
+              elevation: 4.0,
+              borderRadius: BorderRadius.circular(4.0),
+              color: Colors.green[index % 9 * 100],
+              child: Center(
+                child: Text(index.toString()),
+              ),
+            ),
+          );
+        },
       ),
     );
   }