+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.0
 
-* A Widget that allows you to quickly scroll through a list
+* A Widget that helps users quickly scroll through a list
 * Can optionally display a label as the user scrolls
 * 3 styles available:
   * Google Photos style